refactor: remove chatroom context
