--- hooks/context/useChatroom.ts
@@ -1,8 +1,0 @@
-import ChatroomContext from "@/contexts/ChatroomContext";
-import { useContext } from "react";
-
-const useChatroom = () => {
-  return useContext(ChatroomContext);
-};
-
-export default useChatroom;
