+++ package.json
@@ -58,6 +58,7 @@
     "@typescript-eslint/parser": "5.60.0",
     "@types/ws": "8.5.5",
     "autoprefixer": "10.4.14",
+    "bundlewatch": "0.3.3",
     "cross-env": "7.0.3",
     "cypress": "12.15.0",
     "eslint": "8.43.0",
@@ -74,7 +75,6 @@
     "storybook": "7.0.23",
     "tailwindcss": "3.3.2",
     "typescript": "5.1.3",
-    "bundlewatch": "0.3.3",
     "ws": "8.13.0"
   },
   "husky": {
