+++ pages/_app.tsx
@@ -14,7 +14,6 @@ import ApiHistoryProvider from "@/containers/provider/ApiHistoryProvider";
 import ApiHistoryList from "@/components/util/api-history/ApiHistoryList";
 import { Env, getEnv } from "@/lib/env";
 import { ToastQueueProvider } from "@/components/shared/Toast";
-import ChatroomContextProvider from "@/containers/provider/ChatroomProvider";
 
 export type NextPageWithProps<P = unknown, IP = P> = NextPage<P, IP> & {
   getLayout?: (page: ReactElement) => ReactNode;
@@ -48,14 +47,12 @@ export default function App({ Component, pageProps }: AppWithProps) {
       <ModalManager.Provider>
         <AxiosProvider>
           <AuthProvider>
-            <ChatroomContextProvider>
             {getHistory(
               <Startup isAnonymous={isAnonymous}>
                 {getLayout(<Component {...pageProps} />)}
                 {!isProduction && <ApiHistoryList />}
               </Startup>
             )}
-            </ChatroomContextProvider>
           </AuthProvider>
         </AxiosProvider>
       </ModalManager.Provider>
