+++ pages/rooms/[roomId]/index.tsx
@@ -26,6 +26,7 @@ export default function Room() {
     // toggleUserReadyStatus,
     cleanUpRoom,
   } = useRoom();
+
   const { currentUser } = useAuth();
   const { Popup, firePopup } = usePopup();
   const { fetch } = useRequest();
